% Auto-generated: do not edit by hand
\name{myauoDashTextToSpeech}

\alias{myauoDashTextToSpeech}

\title{DashTextToSpeech component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
myauoDashTextToSpeech(id=NULL, paused=NULL, pitch=NULL, rate=NULL, text=NULL,
volume=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{paused}{Logical. The value displayed in the input.}

\item{pitch}{Numeric. The value displayed in the input.}

\item{rate}{Numeric. The value displayed in the input.}

\item{text}{Character. A label that will be printed when this component is rendered.}

\item{volume}{Numeric. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

